src/agent/TicketCreatorAgent/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-mcp-adapters
pydantic-settings
python-dotenv
crewai>=0.80.0
